moved config file with constants into one folder with wdio.conf
