+++ src/specs/card.test.js
@@ -1,5 +1,5 @@
 const BoardsPage = require("../pageobjects/boards.page");
-const { LIST_NAME, BOARD_NAME_CARD, CARD_NAME } = require("../support/config");
+const { LIST_NAME, BOARD_NAME_CARD, CARD_NAME } = require("../configs/config");
 
 describe("Verify action on board:  cards", () => {
   it("Add card to existent list", async function () {
