+++ src/specs/login.test.js
@@ -1,7 +1,7 @@
 const LoginPage = require("../pageobjects/login.page");
 const credentials = require("../configs/wdio.conf");
 const BoardsPage = require("../pageobjects/boards.page");
-const { USER_NAME, EXPECTED_LOGIN_ERROR } = require("../support/config");
+const { USER_NAME, EXPECTED_LOGIN_ERROR } = require("../configs/config");
 
 describe("Verify Login with positive and negative cases", () => {
   it("Login with valid credentials for your free account", async function () {
